Add files via upload
In this game, the user sets up a board that contains a grid of cells, where each cell can be either "living" or "dead".  Once the board is set up and the game is started, it runs itself. The board goes through a sequence of "generations."  In each generation, every cell can change its state from living to dead or vice versa, depending on the number of neighbours that it has.